requirements.txt
requests==2.32.3
beautifulsoup4==4.13.4
pydantic==2.11.5
aiogram>=3.0.0
python-dotenv==1.1.0